requirements.txt
Django
django-filter
djangorestframework
psycopg2-binary
Pillow
python-decouple
django-cors-headers
drf-yasg
gunicorn